requirements.txt
# Drishti Health - Dependencies
# Core
streamlit==1.41.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.19
pydantic==2.10.0

# ML / Data Science
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
xgboost==2.1.3
shap==0.46.0
matplotlib==3.9.3
seaborn==0.13.2

# Computer Vision / Medical Imaging
opencv-python-headless==4.10.0.84
Pillow==11.0.0
ultralytics==8.3.0
torch>=2.1.0
torchvision>=0.16.0
timm==1.0.11

# Voice / NLP
vosk==0.3.45
requests==2.32.3
transformers==4.46.0

# Explainability & Viz
plotly==5.24.0

# Database
aiosqlite==0.20.0

# Utilities
python-dotenv==1.0.1
httpx==0.28.0
